Struct reactor_rt::assembly::DependencyDeclarator

pub struct DependencyDeclarator<'a, 'x, S: ReactorInitializer> { /* fields omitted */ }

Implementations

impl<S: ReactorInitializer> DependencyDeclarator<'_, '_, S>

pub fn declare_triggers(
&mut self,
trigger: TriggerId,
reaction: GlobalReactionId
) -> AssemblyResult<()>

pub fn effects_port<T: Sync>(
&mut self,
reaction: GlobalReactionId,
port: &Port<T>
) -> AssemblyResult<()>

pub fn effects_bank<T: Sync>(
&mut self,
reaction: GlobalReactionId,
port: &PortBank<T>
) -> AssemblyResult<()>

pub fn declare_uses(
&mut self,
reaction: GlobalReactionId,
trigger: TriggerId
) -> AssemblyResult<()>

pub fn bind_ports<T: Sync>(
&mut self,
upstream: &mut Port<T>,
downstream: &mut Port<T>
) -> AssemblyResult<()>

Bind two ports together.

pub fn bind_ports_zip<'a, T: Sync + 'a>(
&mut self,
upstream: impl Iterator<Item = &'a mut Port<T>>,
downstream: impl Iterator<Item = &'a mut Port<T>>
) -> AssemblyResult<()>

Bind the ports of the upstream to those of the downstream, as if zipping both iterators. todo this will just throw away bindings if both iterators are not of the same size normally this should be reported by LFC as a warning, maybe we should implement the same thing here