Module reactor_rt::prelude

The prelude that is imported at the top of reactor files generated by LFC.

Re-exports

pub use crate::Offset::*;

Macros

after

Shorthand for using After together with delay.

assert_tag_is

Convenient macro to assert equality of the current tag. This is just shorthand for using assert_eq! with the syntax of tag.

delay

Creates a Duration value using the same syntax as in LF.

tag

Convenient macro to create a tag. This is just a shorthand for using the constructor together with the syntax of delay.

Structs

AsyncCtx

A type that can affect the logical event queue to implement asynchronous physical actions. This is a “link” to the event system, from the outside world.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

EventTag

The tag of an event.

Instant

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

LogicalAction

A logical action.

Multiport

A multiport is a vector of independent ports (its channels) Multiports have special Lingua Franca syntax, similar to reactor banks.

PhysicalActionRef

A reference to a physical action. This thing is cloneable and can be sent to async threads. The contained action reference is unique and protected by a lock. All operations on the action are

Port

Represents a port, which carries values of type T. Ports reify the data inputs and outputs of a reactor.

ReactionCtx

The context in which a reaction executes. Its API allows mutating the event queue of the scheduler. Only the interactions declared at assembly time are allowed.

Timer

A timer is conceptually a logical action that may re-schedule itself periodically.

Type Definitions

unit

Alias for the unit type, so that it can be written without quotes in LF. Otherwise it needs to be written {= () =}. It is not camel-case as it is actually a primitive type.