Struct reactor_rt::AssemblyCtx

pub struct AssemblyCtx<'x> { /* fields omitted */ }

Helper struct to assemble reactors during initialization. One assembly context is used per reactor, they can’t be shared.

Implementations

impl<'x> AssemblyCtx<'x>

pub fn get_id(&self) -> ReactorId

The ID of the reactor being built.

Panics

If fix_cur_id has not been called.

pub fn fix_cur_id(&mut self) -> ReactorId

Note: this needs to be called after all children reactors have been built, as they’re pushed into the global reactor vec before their parent. So the ID of the parent needs to be fixed only after all descendants have been built.

pub fn new_port<T: Send>(
&mut self,
lf_name: &'static str,
is_input: bool
) -> Port<T>

pub fn new_port_bank_const<T: Send, const N: usize>(
&mut self,
lf_name: &'static str,
is_input: bool
) -> [Port<T>; N]

pub fn new_port_bank<T: Send>(
&mut self,
lf_name: &'static str,
is_input: bool,
len: usize
) -> Vec<Port<T>>

pub fn new_logical_action<T: Send>(
&mut self,
lf_name: &'static str,
min_delay: Option<Duration>
) -> LogicalAction<T>

pub fn new_physical_action<T: Send>(
&mut self,
lf_name: &'static str,
min_delay: Option<Duration>
) -> PhysicalActionRef<T>

pub fn new_timer(
&mut self,
lf_name: &'static str,
offset: Duration,
period: Duration
) -> Timer

pub fn new_reactions<const N: usize>(
&mut self,
num_non_synthetic: usize,
names: [Option<&'static str>; N]
) -> [GlobalReactionId; N]

Create N reactions. The first num_non_synthetic get priority edges, as they are taken to be those declared in LF by the user. The rest do not have priority edges, and their implementation must hence have no observable side-effect.

pub fn declare_triggers(
&mut self,
trigger: TriggerId,
reaction: GlobalReactionId
) -> Result<(), AssemblyError>

pub fn effects_port<T: Send>(
&mut self,
reaction: GlobalReactionId,
port: &Port<T>
) -> Result<(), AssemblyError>

pub fn declare_uses(
&mut self,
reaction: GlobalReactionId,
trigger: TriggerId
) -> Result<(), AssemblyError>

pub fn bind_ports<T: Send>(
&mut self,
upstream: &mut Port<T>,
downstream: &mut Port<T>
) -> Result<(), AssemblyError>

pub fn register_reactor<S: ReactorInitializer + 'static>(&mut self, child: S)

Register a child reactor.

pub fn assemble_sub<S: ReactorInitializer>(
&mut self,
inst_name: &'static str,
args: S::Params
) -> Result<S, AssemblyError>

Assemble a child reactor. The child needs to be registered using Self::register_reactor later.