Module reactor_rt::assembly

Module containing the API to initialize a reactor program.

Structs

AssemblyCtx

Helper struct to assemble reactors during initialization. One assembly context is used per reactor, they can’t be shared.

AssemblyError

An error occurring during initialization of the reactor program. Should never occur unless the graph is built by hand, and not by a Lingua Franca compiler.

AssemblyIntermediate

Intermediate result of assembly.

ComponentCreator

Creates the components of a reactor.

DependencyDeclarator

Declares dependencies between components and reactions.

FinishedReactor

Final result of the assembly of a reactor.

GlobalReactionId

Global identifier for a reaction.

TriggerId

The ID of a trigger component.

Enums

PortKind

Kind of a port.

Traits

ReactorInitializer

Wrapper around the user struct for safe dispatch.

Type Definitions

AssemblyResult