Trait reactor_rt::assembly::ReactorInitializer

pub trait ReactorInitializer: ReactorBehavior {
    type Wrapped;
    type Params;

    const MAX_REACTION_ID: LocalReactionId;

    // Required method
    fn assemble(
        args: Self::Params,
        assembler: AssemblyCtx<'_, Self>
    ) -> AssemblyResult<FinishedReactor<'_, Self>>
       where Self: Sized;
}

Wrapper around the user struct for safe dispatch.

Fields are

the user struct,

ctor parameters of the reactor, and

every logical action and port declared by the reactor.

Required Associated Types

type Wrapped

Type of the user struct, which contains state variables of the reactor. Used by the runtime to produce debug information.

type Params

Type of the construction parameters.

Required Associated Constants

const MAX_REACTION_ID: LocalReactionId

Exclusive maximum value of the local_rid parameter of [ReactorBehavior.react].

Required Methods

fn assemble( args: Self::Params, assembler: AssemblyCtx<'_, Self> ) -> AssemblyResult<FinishedReactor<'_, Self>>
where Self: Sized,

Assemble this reactor. This initializes state variables, produces internal components, assembles children reactor instances, and declares dependencies between them.

Object Safety

This trait is not object safe.