reactor_rt
Module prelude

The prelude that is imported at the top of reactor files generated by LFC.

Re-exports

pub use crate::AsyncCtx;

pub use crate::EventTag;

pub use crate::LogicalAction;

pub use crate::Multiport;

pub use crate::PhysicalActionRef;

pub use crate::Port;

pub use crate::ReactionCtx;

pub use crate::Timer;

pub use crate::Offset::*;

Macros

after

Shorthand for using After together with delay.

assert_tag_is

Convenient macro to assert equality of the current tag. This is just shorthand for using assert_eq! with the syntax of tag.

delay

Creates a Duration value using the same syntax as in LF.

tag

Convenient macro to create a tag. This is just a shorthand for using the constructor together with the syntax of delay.

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

Instant

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

Type Aliases

unit

Alias for the unit type, so that it can be written without quotes in LF. Otherwise it needs to be written {= () =}. It is not camel-case as it is actually a primitive type.