Module reactor_rt::prelude

The prelude that is imported at the top of reactor files generated by LFC.

Re-exports

pub use crate::Offset::*;

pub use crate::after;

pub use crate::assert_tag_is;

pub use crate::delay;

pub use crate::tag;

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

EventTag

The tag of an event.

Instant

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

LogicalAction

A logical action.

PhysicalActionRef

A reference to a physical action. This thing is cloneable and can be sent to async threads. The contained action reference is unique and protected by a lock. All operations on the action are

ReactionCtx

The context in which a reaction executes. Its API allows mutating the event queue of the scheduler. Only the interactions declared at assembly time are allowed.

ReadablePort

A read-only reference to a port.

Timer

A timer is conceptually a logical action that may re-schedule itself periodically.

WritablePort

A write-only reference to a port.